scrollbars: move keyboard handling to root ScrollView item and call forceActiveFocus() on click
This way we can consume events not consumed by children
(if a child has focus, viewportItem.focus will be false, so that wouldn't
be possible if the keyboard was handled inside viewportItem, as previously)

Moreover, I also now forceActiveFocus() on the root ScrollView when the user clicks
inside the ScrollView. This is need because activeFocusOnPress doesn't focus the 
item if *any* of the parents has activeFocusOnPress disabled. 
Messaging app, for instance, has activeFocusOnPress:false in its MainView, and 
that was preventing the ScrollView from getting focus on click.