[scrollbar]: OPTIMIZATION wait for flickableItem initialization. 20% faster benchmark
- Wait for flickableItem Component.onCompleted before doing anything with
it, i.e. reading its properties etc.

Moreover, not waiting for onCompleted meant blocking the thread during the 
instantiation of Flickable for a few msecs (seen with QtCreator profiler),
because the changes in Flickable
properties while it was being initialized were causing logic in Scrollbar to
be fired, hence Flickable could not complete initialization before the 
slots in scrollbar were handled. QtCreator profiler showed 
there was a big chunk of about 5msecs just to "create" the Flickable
because of all the things Scrollbar does when Flickable's properties change
(such as hinting) and the storm of calls to slots that was triggering.

And it was all useless work :) 
The Scrollbar performance benchmark is now 20% faster, just because of 
avoiding useless work :)