Scrollbar: remove leftover margins around the area handling interaction with the thumb
It is a leftover from the previous implementation. 
For a vertical (horizontal) scrollbar the input area of the THUMB is already
as wide (tall) as the TROUGH, so there is some error margin around the
thumb already.

Plus, at the moment these margins are not actually serving a purpose
because the MouseArea is inside a smaller element (the trough Item), 
which prevents it from receiving the mouse events falling in that area