Scrollbar: fix unit tests by passing all parameters to mapToItem. Fixes tests on Qt5.6
mapToItem takes (item, x, y) as parameters. The unit tests were previously
relying on JS automatically defining x,y as 0 when they were not passed
as parameters.  

We were lucky that worked in the past, but Qt5.6 has a stricter
control over parameters and it throws TypeError when not all of them
are provided.

See QTBUG-41686 and its fix.