iconprovider: support getting icon theme name from QIcon and env var
QIcon is supposed to know best which icon theme is currently active.
Query it, so that image://theme/ uri obeys the current icon theme.

As this breaks the tests because they're written with "suru" theme in
mind, also add a way to override the icon theme via the environment
variable and use it in the test.