Port to Qt 5.12
This is a big change to build uitk against newer version of Qt. As in the
past, some new tests now result in crashes or odd failures due to changes
in Qt, so have had to be skipped, though others have been fixed and the
toolkit behaves fine at runtime in testing.